refactor: scafolding based on new pages
