+++ src/pages/App/Daily/index.jsx
@@ -1,6 +1,6 @@
 // components
 import { Heading } from '@chakra-ui/react'
-import { Form } from 'components/App/Home/Form-complete'
+import { Form } from 'components/App/Daily/Form'
 
 export const Daily = () => {
   return (
